+++ src/checkers/wp_plugins/mod.rs
@@ -11,6 +11,7 @@ pub mod elements_ready_lite;
 pub mod gtranslate;
 pub mod jetpack;
 pub mod js_composer;
+pub mod litespeed_cache;
 pub mod really_simple_ssl;
 pub mod revslider;
 pub mod woocommerce;